+++ src/venmo/shared/errors.js
@@ -1,5 +1,21 @@
 'use strict';
 
+/**
+ * @name BraintreeError.Venmo - Creation Error Codes
+ * @description Errors that occur when [creating the Venmo component](/current/module-braintree-web_venmo.html#.create).
+ * @property {MERCHANT} VENMO_NOT_ENABLED Occurs when Venmo is not enabled on the Braintree control panel.
+ * @property {MERCHANT} VENMO_INVALID_PROFILE_ID Occurs when Venmo is intilaized with a profile id, but it is invalid.
+ */
+
+/**
+ * @name BraintreeError.Venmo - tokenize Error Codes
+ * @description Errors that occur when using the [`tokenize` method](/current/Venmo.html#tokenize).
+ * @property {MERCHANT} VENMO_TOKENIZATION_REQUEST_ACTIVE Occurs when `tokenize` is called when the flow is already in progress.
+ * @property {UNKNOWN} VENMO_APP_FAILED Occurs when tokenization fails.
+ * @property {CUSTOMER} VENMO_APP_CANCELED Occurs when customer cancels flow from the Venmo app.
+ * @property {CUSTOMER} VENMO_CANCELED Occurs when customer cancels the flow or Venmo app is not available.
+ */
+
 var BraintreeError = require('../../lib/braintree-error');
 
 module.exports = {